refactor: address review feedback on toast module
- Move errorLogWithToast into its own module to break the
  utils -> toast -> utils import cycle, so importing cn no longer
  pulls in the toast manager
- Widen ToastType with (string & {}) to allow custom toast types
  while keeping literal autocomplete, and check icon lookup with
  Object.hasOwn instead of the prototype-chain-sensitive in operator
