+++ src/lib/background-image/use-background-image.ts
@@ -1,7 +1,7 @@
 import { useBackgroundImageFileDb } from "@/lib/file-db/file-db-store";
 import { useCallback } from "react";
 import { toast } from "@/components/ui/toast";
-import { errorLogWithToast } from "../utils";
+import { errorLogWithToast } from "@/lib/error-toast";
 
 export function useBackgroundImage() {
   const {
