+++ src/lib/media-compositor/use-recorder.ts
@@ -7,7 +7,7 @@ import {
 import { toast } from "@/components/ui/toast";
 import { runRecorder } from "./run-recorder-worker";
 import type { ActivePhase } from "./export-progress-tracker";
-import { errorLogWithToast } from "../utils";
+import { errorLogWithToast } from "@/lib/error-toast";
 import type { MidiTracks } from "@/lib/midi/midi";
 import type { AudioSource } from "@/lib/audio/audio";
 import type { RendererConfig } from "@/lib/renderers/renderer";
