+++ src/lib/utils.ts
@@ -1,7 +1,6 @@
 import { clsx, type ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
 import { closeDb } from "@/lib/file-db/file-db";
-import { toast } from "@/components/ui/toast";
 
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
@@ -50,10 +49,4 @@ export function startViewTransition(callback: () => void, options?: { types?: st
     },
     ...options,
   });
-}
-
-export function errorLogWithToast(message: string, error?: unknown) {
-  console.error(...[message, error].filter(Boolean));
-  const description = error instanceof Error ? error.message : undefined;
-  toast.add({ title: message, description, type: "error" });
 }
